UML class relation diagram is now separate images for each graph
