+++ pystruct/html_utils/templates/dropmenu.html
