+++ pystruct/html_utils/templates/image.html
